blogger-gui/requirements.txt
google-api-python-client==2.105.0
google-auth-oauthlib==1.1.0
PyQt6==6.6.1
requests==2.31.0
pytest==7.4.3
